FIX: msvc static analysis warnings
- _BitScanReverse takes unsigned long, not uint32_t
- disable C4214:  nonstandard extension used: bit field types other than int

+++ lib/src/bits.h
@@ -26,7 +26,7 @@ static inline uint32_t count_leading_zeros(uint32_t x) {
 
 static inline uint32_t count_leading_zeros(uint32_t x) {
   if (x == 0) return 32;
-  uint32_t result;
+  unsigned long result;
   _BitScanReverse(&result, x);
   return 31 - result;
 }
